% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empirical_data.R
\docType{data}
\name{forts}
\alias{forts}
\title{Forts Location}
\format{Data frame with columns
\describe{
 \item{id}
 \item{type}{1s are forts and NAs are gyms}
 \item{latitude}
 \item{longitude}
 \item{enabled}
 \item{last_modified_timestamp_ms}
 \item{cooldown_complete_timestamp_ms}
 \item{active_fort_modifier}
 \item{owned_by_team}
 \item{guard_pokemon_id}
 \item{gym_points}
 \item{is_in_battle}
}}
\source{
This package
}
\usage{
forts
}
\description{
Latitudes and longitudes of forts and gyms in Tainan.
}
\examples{
  forts
}
\keyword{datasets}

