% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bot_data.R
\docType{data}
\name{items}
\alias{items}
\title{Items in Pokemon Go}
\format{Data frame with columns
\describe{
 \item{id}{join this id with data dumped from the game}
 \item{item}
}}
\source{
\url{https://github.com/PokemonGoF/PokemonGo-Bot}, downloaded 2016-08-24
}
\usage{
items
}
\description{
Items Ids and Names
}
\examples{
  items
}
\keyword{datasets}

