% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bot_data.R
\docType{data}
\name{level_to_cpm}
\alias{level_to_cpm}
\title{Level To Combat Point Multiplier(cpm)}
\format{Data frame with columns
\describe{
 \item{level}{trainer's level}
 \item{cpm}{Combat Point Multiplier}
}}
\source{
\url{https://github.com/PokemonGoF/PokemonGo-Bot}, downloaded 2016-08-24
}
\usage{
level_to_cpm
}
\description{
Convert Pokemon level to Combat Point Multiplier(cpm)
}
\examples{
  level_to_cpm
}
\keyword{datasets}

