% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bot_data.R
\docType{data}
\name{moves}
\alias{moves}
\title{Pokemon Go Moves}
\format{Data frame with columns
\describe{
 \item{id}{in game id}
 \item{name}
 \item{type}
 \item{damage}
 \item{duration}
 \item{energy}{fast moves charge energy, while charged moves consume energy}
 \item{dps}{damage per second}
 \item{move}{fast or charged move}
}}
\source{
\url{https://github.com/PokemonGoF/PokemonGo-Bot}, downloaded 2016-08-24
}
\usage{
moves
}
\description{
Fast and charged moves of Pokemon
}
\examples{
  moves
}
\keyword{datasets}

