% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pokemon_names.R
\docType{data}
\name{pokemon_names}
\alias{pokemon_names}
\title{Pokedex Names.}
\format{Data frame with columns
\describe{
 \item{index}{Integer Pokedex index of the pokemon}
 \item{name}{Proper name of pokemon}
}}
\source{
\url{https://gist.github.com/noxwyll/8c821408b1d943c3e500d91d77266368/f264e42f2e83bbe845841717f74b276c9801ba7d}, downloaded 2016-08-19
}
\usage{
pokemon_names
}
\description{
Pokedex index and names of the first generation 151 pokemons.
}
\examples{
pokemon_names
}
\keyword{datasets}

