% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bot_data.R
\docType{data}
\name{pokemon_names_zh_tw}
\alias{pokemon_names_zh_tw}
\title{Pokemon Names in Taiwan}
\format{Data frame with columns
\describe{
 \item{index}
 \item{en_name}
 \item{zh_tw_name}
}}
\source{
\url{https://github.com/PokemonGoF/PokemonGo-Bot}, downloaded 2016-08-24
}
\usage{
pokemon_names_zh_tw
}
\description{
Pokemon Names in Taiwan
}
\examples{
  pokemon_names_zh_tw
}
\keyword{datasets}

