% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empirical_data.R
\docType{data}
\name{pokemon_subjects}
\alias{pokemon_subjects}
\title{Pokemon Subjects Collected by Trainers}
\format{Data frame with columns
\describe{
 \item{trainer}{name of the trainer}
 \item{id}{id of pokemon}
 \item{creation_time_ms}
 \item{pokemon_id}
 \item{captured_cell_id}
 \item{pokeball}
 \item{height_m}
 \item{weight_kg}
 ...
}}
\source{
This package
}
\usage{
pokemon_subjects
}
\description{
Yes, conducting pokemon studies without transfering
all your pokemons to a doctor is possible. The downside
is you are not rewarded with a candy.
}
\examples{
  pokemon_subjects
}
\keyword{datasets}

